refactor: improve share feedback
