+++ src/components/Profile/ShareFeedback.tsx
@@ -6,7 +6,7 @@ import { router } from "expo-router";
 
 const ShareFeedback = () => {
   const navigateToFeedback = () => {
-    router.push("/feedback");
+    router.push("/(profile)/share-feedback");
   };
 
   return (
